Refactor Navigator to have a class wrapper around the GPS and compass
Since Navigator is a module, rather than a library, it's not really a
bad thing to allow direct access through it to the GPS and compass
devices. As such, Navigator doesn't need to implement wrapper functions.
Access to the GPS and compass devices outside of Navigator is provided
by the gps() and mag() functions, so the only real change to clients is

- gps.method();
- compass.method();
+ nav.gps()->method();
+ nav.mag()->method();